# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: OpenADR 3 API
  description: |
    The OpenADR 3 API supports energy retailer to energy customer Demand Response programs.
    See OpenADR 3 User Guide and Defintions for detailed descriptions of usage.
    The API includes the following capabilities and operations:

    __Manage programs:__

    * Create/Update/Delete a program
    * Search programs

    __Manage events:__

    * Create/Update/Delete an event
    * Search events

    __Manage reports:__

    * Create/Update/Delete a report
    * Search reports

    __Manage subscriptions:__

    * Create/Update/Delete subscriptions to programs, events, and reports
    * Search subscriptions
    * Subscriptions allows clients to register a callback URL (webhook) to be notified
      on the change of state of a resource

    __Manage vens:__

    * Create/Update/Delete vens and ven resources
    * Search ven and ven resources

    __Manage tokens:__

    * Obtain an access token
    * This endpoint is provided as a convenience and may be neglected in a commercial implementation
  contact:
    email: frank@pajaritotech.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.1.0
servers:
- url: http://localhost:8081/openadr3/3.1.0
  description: base path
paths:
...
  /subscriptions:
    get:
      tags:
      - subscriptions
      summary: search subscriptions
      description: |
        List all subscriptions.
        May filter results by programID and clientName as query params.
        May filter results by targetType and targetValues as query params.
        May filter results by objects as query param. See objectTypes schema.
        Use skip and pagination query params to limit response size.
      operationId: search_subscriptions
      parameters:
      - name: programID
        in: query
        description: filter results to subscriptions with programID.
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/objectID'
      - name: clientName
        in: query
        description: filter results to subscriptions with clientName.
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetType
        in: query
        description: "Indicates targeting type, e.g. GROUP"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetValues
        in: query
        description: "List of target values, e.g. group names"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: objects
        in: query
        description: list of objects to subscribe to.
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            $ref: '#/components/schemas/objectTypes'
      - name: skip
        in: query
        description: number of records to skip for pagination.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return.
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/subscription'
                x-content-type: application/json
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.subscriptions_controller
    post:
      tags:
      - subscriptions
      summary: create subscription
      description: Create a new subscription.
      operationId: create_subscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/subscription_request'
        required: true
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if identical object
            already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      callbacks:
        notifyEvent:
          '{$request.body#/callbackUrl}':
            post:
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/notification'
                required: true
              responses:
                "200":
                  description: Your server returns this code if it accepts the callback.
      security:
      - oAuth2ClientCredentials:
        - write_subscriptions
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.subscriptions_controller
  /subscriptions/{subscriptionID}:
    get:
      tags:
      - subscriptions
      summary: search subscriptions by ID
      description: Return the subscription specified by subscriptionID specified in
        path.
      operationId: search_subscription_by_id
      parameters:
      - name: subscriptionID
        in: path
        description: object ID of the associated subscription.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.subscriptions_controller
    put:
      tags:
      - subscriptions
      summary: update  subscription
      description: Update the subscription specified by subscriptionID specified in
        path.
      operationId: update_subscription
      parameters:
      - name: subscriptionID
        in: path
        description: object ID of the associated subscription.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      requestBody:
        description: subscription item to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/subscription_request'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if program with
            the same programName exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_subscriptions
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.subscriptions_controller
    delete:
      tags:
      - subscriptions
      summary: delete  subscription
      description: Delete the subscription specified by subscriptionID specified in
        path.
      operationId: delete_subscription
      parameters:
      - name: subscriptionID
        in: path
        description: object ID of the associated subscription.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_subscriptions
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.subscriptions_controller
...
components:
  schemas:
    program:
      type: object
      description: Server provided representation of program
      allOf:
      - $ref: '#/components/schemas/object_metadata'
      - $ref: '#/components/schemas/program_request'
...
    report:
      type: object
      description: Server provided representation of report
      allOf:
      - $ref: '#/components/schemas/object_metadata'
      - $ref: '#/components/schemas/report_request'
    report_request:
      required:
      - clientName
      - eventID
      - programID
      - resources
      type: object
      properties:
        programID:
          $ref: '#/components/schemas/objectID'
        eventID:
          $ref: '#/components/schemas/objectID'
        clientName:
          maxLength: 128
          minLength: 1
          type: string
          description: User generated identifier; may be VEN ID provisioned out-of-band.
          example: VEN-999
        reportName:
          type: string
          description: User defined string for use in debugging or User Interface.
          nullable: true
          example: Battery_usage_04112023
        payloadDescriptors:
          type: array
          description: A list of reportPayloadDescriptors.
          nullable: true
          items:
            $ref: '#/components/schemas/reportPayloadDescriptor'
        resources:
          type: array
          description: A list of objects containing report data for a set of resources.
          items:
            $ref: '#/components/schemas/report_request_resources'
      description: report object.
    event:
      type: object
      description: Server provided representation of event
      allOf:
      - $ref: '#/components/schemas/object_metadata'
      - $ref: '#/components/schemas/event_request'
...
    subscription:
      type: object
      description: Server provided representation of subscription
      allOf:
      - $ref: '#/components/schemas/object_metadata'
      - $ref: '#/components/schemas/subscription_request'
    subscription_request:
      required:
      - clientName
      - objectOperations
      - programID
      type: object
      properties:
        clientName:
          maxLength: 128
          minLength: 1
          type: string
          description: "User generated identifier, may be VEN identifier provisioned\
            \ out-of-band."
          example: VEN-999
        programID:
          $ref: '#/components/schemas/objectID'
        objectOperations:
          type: array
          description: list of objects and operations to subscribe to.
          items:
            $ref: '#/components/schemas/subscription_request_objectOperations'
        targets:
          type: array
          description: A list of valuesMap objects. Used by server to filter callbacks.
          nullable: true
          items:
            $ref: '#/components/schemas/valuesMap'
      description: |
        An object created by a client to receive notification of operations on objects.
        Clients may subscribe to be notified when a type of object is created,
        updated, or deleted.
    ven:
      type: object
      description: Server provided representation of ven
      allOf:
      - $ref: '#/components/schemas/object_metadata'
      - $ref: '#/components/schemas/ven_request'
...
    resource:
      type: object
      description: Server provided representation of resource
      allOf:
      - $ref: '#/components/schemas/object_metadata'
      - $ref: '#/components/schemas/resource_request'
...
    object_metadata:
      required:
      - createdDateTime
      - id
      - modificationDateTime
      - objectType
      type: object
      properties:
        id:
          $ref: '#/components/schemas/objectID'
        createdDateTime:
          $ref: '#/components/schemas/dateTime'
        modificationDateTime:
          $ref: '#/components/schemas/dateTime'
        objectType:
          $ref: '#/components/schemas/objectTypes'
      description: metadata common to all addressable objects. Values provided by
        VTN on object creation.
...
    valuesMap:
      required:
      - type
      - values
      type: object
      properties:
        type:
          maxLength: 128
          minLength: 1
          type: string
          description: |
            Enumerated or private string signifying the nature of values.
            E.G. "PRICE" indicates value is to be interpreted as a currency.
          example: PRICE
        values:
          type: array
          description: A list of data points. Most often a singular value such as
            a price.
          example:
          - 0.17
          items:
            anyOf:
            - type: number
            - type: integer
            - type: string
            - type: boolean
            - $ref: '#/components/schemas/point'
      description: |
        Represents one or more values associated with a type.
        E.g. a type of PRICE contains a single float value.
      example:
        values:
        - 0.17
        type: PRICE
    point:
      required:
      - x
      - "y"
      type: object
      properties:
        x:
          type: number
          description: A value on an x axis.
          format: float
          example: 1
        "y":
          type: number
          description: A value on a y axis.
          format: float
          example: 2
      description: A pair of floats typically used as a point on a 2 dimensional grid.
...
    objectID:
      maxLength: 128
      minLength: 1
      pattern: "^[a-zA-Z0-9_-]*$"
      type: string
      description: URL safe VTN assigned object ID.
      example: object-999
    notification:
      required:
      - object
      - objectType
      - operation
      type: object
      properties:
        objectType:
          $ref: '#/components/schemas/objectTypes'
        operation:
          type: string
          description: the operation on on object that triggered the notification.
          example: CREATE
          enum:
          - CREATE
          - READ
          - UPDATE
          - DELETE
        object:
          type: object
          description: the object that is the subject of the notification.
          example: {}
          discriminator:
            propertyName: objectType
          oneOf:
          - $ref: '#/components/schemas/program'
          - $ref: '#/components/schemas/report'
          - $ref: '#/components/schemas/event'
          - $ref: '#/components/schemas/subscription'
          - $ref: '#/components/schemas/ven'
          - $ref: '#/components/schemas/resource'
        targets:
          type: array
          description: A list of valuesMap objects.
          nullable: true
          items:
            $ref: '#/components/schemas/valuesMap'
      description: |
        VTN generated object included in request to subscription callbackUrl.
    objectTypes:
      type: string
      description: Types of objects addressable through API.
      example: EVENT
      enum:
      - PROGRAM
      - EVENT
      - REPORT
      - SUBSCRIPTION
      - VEN
      - RESOURCE
    dateTime:
      type: string
      description: datetime in ISO 8601 format
      format: date-time
      example: 2023-06-15T09:30:00Z
...
    problem:
      type: object
      properties:
        type:
          type: string
          description: |
            An absolute URI that identifies the problem type.
            When dereferenced, it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          format: uri
          example: https://zalando.github.io/problem/constraint-violation
          default: about:blank
        title:
          type: string
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable.
        status:
          maximum: 600
          exclusiveMaximum: true
          minimum: 100
          type: integer
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          format: int32
          example: 503
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
        instance:
          type: string
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
          format: uri
      description: |
        reusable error response. From https://opensource.zalando.com/problem/schema.yaml.
...
    subscription_request_objectOperations:
      required:
      - callbackUrl
      - objects
      - operations
      type: object
      properties:
        objects:
          type: array
          description: list of objects to subscribe to.
          items:
            $ref: '#/components/schemas/objectTypes'
        operations:
          type: array
          description: list of operations to subscribe to.
          items:
            type: string
            description: object operation to subscribe to.
            example: CREATE
            enum:
            - READ
            - CREATE
            - UPDATE
            - DELETE
        callbackUrl:
          type: string
          description: User provided webhook URL.
          format: uri
          example: https://myserver.com/send/callback/here
        bearerToken:
          type: string
          description: |
            User provided token.
            To avoid custom integrations, callback endpoints
            should accept the provided bearer token to authenticate VTN requests.
          nullable: true
          example: NCEJGI9E8ER9802UT9HUG
      description: "object type, operations, and callbackUrl."
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      description: Client credential flow.
      flows:
        clientCredentials:
          tokenUrl: auth/token
          scopes:
            read_all: VENs and BL can read all resources
            write_programs: Only BL can write to programs
            write_events: Only BL can write to events
            write_reports: only VENs can write to reports
            write_subscriptions: VENs and BL can write to subscriptions
            write_vens: VENS and BL can write to vens and resources
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_oAuth2ClientCredentials
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_oAuth2ClientCredentials
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth